refactor: move PageNavigator component to common directory for better organization
