+++ frontend/src/app/profile/[username]/favourites/page.tsx
@@ -11,7 +11,7 @@ import type { User } from '@/lib/profile/types'
 import type { Media } from '@/lib/media/types'
 import { getUserProfile, getUserFavs } from '@/lib/profile/api'
 import MediaCard from '@/components/media/MediaCard'
-import PageNavigator from '@/components/PageNavigator'
+import PageNavigator from '@/components/common/PageNavigator'
 
 const FavouritesPage = () => {
   const { username } = useParams() as { username: string }
